+++ src/main/java/Listener.java
@@ -1,0 +1,9 @@
+import org.testng.ITestListener;
+import org.testng.ITestResult;
+
+public class Listener implements ITestListener {
+
+    public void onTestFailure(ITestResult result) {
+        System.out.println("Make a screenshot...");
+    }
+}
