+++ src/main/java/Retry.java
@@ -1,0 +1,25 @@
+import org.testng.IRetryAnalyzer;
+import org.testng.ITestResult;
+
+public class Retry implements IRetryAnalyzer {
+    private static final int MAX_RETRY = 5;
+    private int attempt = 1;
+
+    @Override
+    public boolean retry(ITestResult iTestResult) {
+        if (!iTestResult.isSuccess()) {
+            if (attempt < MAX_RETRY) {
+                attempt++;
+                iTestResult.setStatus(ITestResult.FAILURE);
+                System.out.println("Retrying one more time...");
+                return true;
+            } else {
+                iTestResult.setStatus(ITestResult.FAILURE);
+            }
+        } else {
+            iTestResult.setStatus(ITestResult.SUCCESS);
+        }
+
+        return false;
+    }
+}
