+++ src/main/java/core/WaitsService.java
@@ -1,0 +1,85 @@
+package core;
+
+import configuration.ReadProperties;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedCondition;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.FluentWait;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+import java.util.List;
+import java.util.NoSuchElementException;
+
+public class WaitsService {
+    private WebDriver driver;
+    private Duration timeout;
+    private WebDriverWait wait;
+
+    public WaitsService(WebDriver driver, Duration timeout) {
+        this.driver = driver;
+        this.timeout = timeout;
+        wait = new WebDriverWait(driver, timeout);
+    }
+
+    public WaitsService(WebDriver driver) {
+        this.driver = driver;
+        this.timeout = Duration.ofSeconds(ReadProperties.timeout());
+        wait = new WebDriverWait(driver, this.timeout);
+    }
+
+    public WebElement waitForExists(By by) {
+        return wait.until(ExpectedConditions.presenceOfElementLocated(by));
+    }
+
+    public WebElement waitForVisibilityLocatedBy(By by) {
+        return wait.until(ExpectedConditions.visibilityOfElementLocated(by));
+    }
+
+    public WebElement waitForElementClickable(By by) {
+        return wait.until(ExpectedConditions.elementToBeClickable(by));
+    }
+
+    public boolean waitForElementInvisible(WebElement webElement) {
+        return wait.until(ExpectedConditions.invisibilityOf(webElement));
+    }
+
+    public List<WebElement> waitForAllVisibleElementsLocatedBy(By locator) {
+        return wait.until(ExpectedConditions.visibilityOfAllElementsLocatedBy(locator));
+    }
+
+    public WebElement fluentWaitForElement(By by) {
+        FluentWait<WebDriver> fluent = new FluentWait<>(driver)
+                .withTimeout(Duration.ofSeconds(30))
+                .pollingEvery(Duration.ofMillis(50))
+                .ignoring(NoSuchElementException.class);
+
+        return fluent.until(driver -> driver.findElement(by));
+    }
+
+    public Boolean fluentWaitForInvisibility(By by) {
+        FluentWait<WebDriver> fluent = new FluentWait<>(driver)
+                .withTimeout(Duration.ofSeconds(30))
+                .pollingEvery(Duration.ofMillis(50))
+                .ignoring(RuntimeException.class);
+
+        return (Boolean) fluent.until(driver -> driver.findElements(by).size() == 0 ? true : new RuntimeException());
+    }
+
+    public boolean waitForInvisibility(WebElement element) {
+        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        return wait.until(ExpectedConditions.invisibilityOf(element));
+    }
+
+    public void waitForTextToBePresentInElementLocated(By locator, String text) {
+        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        wait.until(ExpectedConditions.textToBePresentInElementLocated(locator, text));
+    }
+
+    public void until(ExpectedCondition<Boolean> condition) {
+        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+        wait.until(condition);
+    }
+}
