--- src/main/java/services/BrowserService.java
@@ -1,75 +1,0 @@
-package services;
-
-import configuration.ReadProperties;
-import io.github.bonigarcia.wdm.WebDriverManager;
-import io.github.bonigarcia.wdm.config.DriverManagerType;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.edge.EdgeDriver;
-import org.openqa.selenium.edge.EdgeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxOptions;
-
-public class BrowserService {
-    private WebDriver driver = null;
-    private DriverManagerType driverManagerType;
-
-    public BrowserService() {
-        switch (ReadProperties.browserName().toLowerCase()) {
-            case "chrome":
-                driverManagerType = DriverManagerType.CHROME;
-                WebDriverManager.getInstance(driverManagerType).setup(); // clearDriverCache(). перед setup()
-                driver = new ChromeDriver(getChromeOptions());
-                break;
-            case "firefox":
-                driverManagerType = DriverManagerType.FIREFOX;
-                WebDriverManager.getInstance(driverManagerType).setup(); // clearDriverCache(). перед setup()
-                driver = new FirefoxDriver(getFirefoxOptions());
-                break;
-            case "edge":
-                driverManagerType = DriverManagerType.EDGE;
-                WebDriverManager.getInstance(driverManagerType).setup(); // clearDriverCache(). перед setup()
-                driver = new EdgeDriver(getEdgeOptions());
-                break;
-            default:
-                System.out.println("Browser " + ReadProperties.browserName() + " is not supported.");
-                break;
-        }
-    }
-
-    public WebDriver getDriver() {
-        //driver.manage().timeouts().implicitlyWait()
-        //driver.manage().timeouts().pageLoadTimeout()
-        //driver.manage().timeouts().scriptTimeout()
-
-        return driver;
-    }
-
-    private ChromeOptions getChromeOptions() {
-        ChromeOptions chromeOptions = new ChromeOptions();
-        chromeOptions.addArguments("--disable-gpu");
-        //chromeOptions.addArguments("--window-size=1920,1200");
-        chromeOptions.addArguments("--ignore-certificate-errors");
-        chromeOptions.addArguments("--silent");
-        chromeOptions.addArguments("--start-maximized");
-        //chromeOptions.addArguments("--headless");
-        //chromeOptions.addArguments("--incognito");
-
-        return chromeOptions;
-    }
-
-    private FirefoxOptions getFirefoxOptions() {
-        FirefoxOptions firefoxOptions = new FirefoxOptions();
-        firefoxOptions.addArguments();
-
-        return firefoxOptions;
-    }
-
-    private EdgeOptions getEdgeOptions() {
-        EdgeOptions edgeOptions = new EdgeOptions();
-        edgeOptions.addArguments();
-
-        return edgeOptions;
-    }
-}
