+++ src/test/java/_otherPackages/baseEntities/BaseTest.java
@@ -1,0 +1,30 @@
+package _otherPackages.baseEntities;
+
+import configuration.ReadProperties;
+import org.openqa.selenium.WebDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import core.BrowserService;
+import _otherPackages.steps.LoginStep;
+import _otherPackages.steps.NavigationStep;
+
+public class BaseTest {
+    protected WebDriver driver;
+    protected LoginStep loginStep;
+    protected NavigationStep navigationStep;
+
+    @BeforeMethod
+    public void setup() {
+        driver = new BrowserService().getDriver();
+
+        loginStep = new LoginStep(driver);
+        navigationStep = new NavigationStep(driver);
+
+        driver.get(ReadProperties.getUrl());
+    }
+
+    @AfterMethod
+    public void tearDown() {
+        driver.quit();
+    }
+}
