+++ src/test/java/_otherPackages/pages/DashboardPage.java
@@ -1,0 +1,35 @@
+package _otherPackages.pages;
+
+import _otherPackages.baseEntities.BasePage;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+
+public class DashboardPage extends BasePage {
+    private final static String pagePath = "/index.php?/dashboard";
+
+    public TopMenuPage topMenuPage;
+
+    // Блок описания селекторов для элементов
+    private By headerTitleLabelLocator = By.xpath("//div[contains(@class, 'content-header-title') and contains(text(), 'All Projects')]");
+
+    public DashboardPage(WebDriver driver) {
+        super(driver);
+
+        topMenuPage = new TopMenuPage(driver);
+    }
+
+    @Override
+    protected By getPageIdentifier() {
+        return headerTitleLabelLocator;
+    }
+
+    public void openPageByUrl() {
+        super.openPageByUrl(pagePath);
+    }
+
+    // Блок атомарных методов
+    public WebElement getHeaderTitleLabel() {
+        return waitsService.waitForExists(headerTitleLabelLocator);
+    }
+}
