+++ src/test/java/_otherPackages/pages/ProjectsPage.java
@@ -1,0 +1,28 @@
+package _otherPackages.pages;
+
+import _otherPackages.baseEntities.BasePage;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class ProjectsPage extends BasePage {
+    private final static String pagePath = "/index.php?/admin/projects/overview";
+
+    public SideMenuPage sideMenuPage;
+
+    private By headerTitleLabelLocator = By.xpath("//div[contains(@class, 'content-header-title') and contains(text(), 'Projects')]");
+
+    public ProjectsPage(WebDriver driver) {
+        super(driver);
+
+        sideMenuPage = new SideMenuPage(driver);
+    }
+
+    @Override
+    protected By getPageIdentifier() {
+        return headerTitleLabelLocator;
+    }
+
+    public void openPageByUrl() {
+        super.openPageByUrl(pagePath);
+    }
+}
