+++ src/test/java/_otherPackages/pages/TopMenuPage.java
@@ -1,0 +1,18 @@
+package _otherPackages.pages;
+
+import _otherPackages.baseEntities.BasePage;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class TopMenuPage extends BasePage {
+    private By topPanelLocator = By.className("top-panel");
+
+    public TopMenuPage(WebDriver driver) {
+        super(driver);
+    }
+
+    @Override
+    protected By getPageIdentifier() {
+        return topPanelLocator;
+    }
+}
