+++ src/test/java/_otherPackages/steps/LoginStep.java
@@ -1,0 +1,36 @@
+package _otherPackages.steps;
+
+import _otherPackages.baseEntities.BaseStep;
+import org.openqa.selenium.WebDriver;
+import _otherPackages.pages.DashboardPage;
+import _otherPackages.pages.LoginPage;
+
+public class LoginStep extends BaseStep {
+
+    public LoginStep(WebDriver driver) {
+        super(driver);
+    }
+
+    public DashboardPage successLogin(String email, String psw) {
+        login(email, psw);
+
+        return dashboardPage;
+    }
+
+    public LoginPage incorrectLogin(String email, String psw) {
+        login(email, psw);
+
+        return loginPage;
+    }
+
+    private void login(String email, String psw) {
+        loginPage.getEmailInput().sendKeys(email);
+        loginPage.getPswInput().sendKeys(psw);
+        loginPage.getLogInButton().click();
+    }
+
+    public LoginPage logout() {
+
+        return loginPage;
+    }
+}
