+++ src/test/java/_otherPackages/tests/TestCaseTests.java
@@ -1,0 +1,17 @@
+package _otherPackages.tests;
+
+import _otherPackages.baseEntities.BaseTest;
+import org.testng.annotations.Test;
+
+public class TestCaseTests extends BaseTest {
+
+    @Test
+    public void successfulTestCaseValidationTest() {
+        // 1. Войти в приложение
+        // 2. Выбрать проект
+        // 3. Перейти на закладку Test Cases
+        // 4. Добавить валидный Test Case
+
+
+    }
+}
