--- src/test/java/baseEntities/BaseTest.java
@@ -1,30 +1,0 @@
-package baseEntities;
-
-import configuration.ReadProperties;
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import services.BrowserService;
-import steps.LoginStep;
-import steps.NavigationStep;
-
-public class BaseTest {
-    protected WebDriver driver;
-    protected LoginStep loginStep;
-    protected NavigationStep navigationStep;
-
-    @BeforeMethod
-    public void setup() {
-        driver = new BrowserService().getDriver();
-
-        loginStep = new LoginStep(driver);
-        navigationStep = new NavigationStep(driver);
-
-        driver.get(ReadProperties.getUrl());
-    }
-
-    @AfterMethod
-    public void tearDown() {
-        driver.quit();
-    }
-}
