--- src/test/java/pages/SideMenuPage.java
@@ -1,18 +1,0 @@
-package pages;
-
-import baseEntities.BasePage;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-
-public class SideMenuPage extends BasePage {
-    private By sidebarLocator = By.id("sidebar");
-
-    public SideMenuPage(WebDriver driver) {
-        super(driver);
-    }
-
-    @Override
-    protected By getPageIdentifier() {
-        return sidebarLocator;
-    }
-}
