--- src/test/java/steps/NavigationStep.java
@@ -1,17 +1,0 @@
-package steps;
-
-import baseEntities.BaseStep;
-import org.openqa.selenium.WebDriver;
-import pages.ProjectsPage;
-
-public class NavigationStep extends BaseStep {
-
-    public NavigationStep(WebDriver driver) {
-        super(driver);
-    }
-
-    public ProjectsPage navigateToProjectsPage() {
-        projectsPage.openPageByUrl();
-        return projectsPage;
-    }
-}
