--- src/test/java/tests/TestCaseTests.java
@@ -1,17 +1,0 @@
-package tests;
-
-import baseEntities.BaseTest;
-import org.testng.annotations.Test;
-
-public class TestCaseTests extends BaseTest {
-
-    @Test
-    public void successfulTestCaseValidationTest() {
-        // 1. Войти в приложение
-        // 2. Выбрать проект
-        // 3. Перейти на закладку Test Cases
-        // 4. Добавить валидный Test Case
-
-
-    }
-}
